requirements.txt
cv_bridge==1.13.0.post0
einops==0.4.1
keyboard==0.13.5
matplotlib==3.1.1
numpy==1.19.5
opencv_python==4.1.2.30
pandas==0.25.3
Pillow==11.2.1
pipython==2.11.0.6
pytransform3d==1.2
scikit_image==0.16.2
scipy==1.3.2
skimage==0.0
torch==1.10.2
torchvision==0.11.3
ultralytics==8.3.156
